Half-decent code generation!
LLVM generates object code, leaning heavily on the runtime for
support.

The UI is still a WIP; for now, the frontend only supports dumping
the LLVM IR to stderr.

The main things missing for basic correctness are:

- Handling of function arguments; currently the caller must use
  the right number of arguments.
- Handling non-local scoping.
- Make and/or non-greedy (oops - this was an oversight from just
  blindly copying operations across to the runtime).

Beyond that, there's no GC yet so it's easy to engineer an OOM.
However, it already beats some transpiling implementations, to go
by a bit of informal benchmarking.