new: Add support for executing closures in codegen
This is a fairly extensive change to the IR and LLVM pass. A user-
defined function is now a block with the arity and closure captures
as well as the function pointer. All closure variables captured in
a function are placed in a heap-allocated array which is passed to
all functions which (transitively) use it. When a function is called,
it is passed the pointer to its closure block. For functions not
taking a closure, this is an i8* pointing to a zero value; I could
do some extra work to make it a null pointer instead.

As well as tweaking the IR representations, I changed function calls
to switch over the arity at the call site rather than in a shim. This
feels like less of a hack but is still pretty ugly. I may iterate
on it more in future.